Refactor collection proxies
