+++ src/NodeApi/JSArray.Enumerator.cs
