+++ src/NodeApi/JSIterable.Enumerator.cs
