+++ src/NodeApi/JSMap.Enumerator.cs
